refactor: extract grid layout hook and adjust display rotation to scale with paginated content
